Add translation keys to locales for admins
- if user is an admin, add [ids] to selectable locales
  - when selected, the id of the text in translations is displayed instead of the translated text
  - helps admins to identify which translation to edit to modify some text on the website
- also some cleanup
  - remove dead code
  - remove console logging
  - use typescript/runes in all components, add missing typedefs
  - inline Frontpage component into /page
  - use `user` store to login and update current user data, remove no longer used utils/login.ts
