+++ frontend/src/lib/components/WhiteCircle.svelte
@@ -1,4 +1,6 @@
-<script>
+<svelte:options runes={true} />
+
+<script lang="ts">
 let { solid = false } = $props();
 </script>
