+++ frontend/src/lib/i18n.svelte.ts
@@ -1,4 +1,5 @@
 import { getLanguages } from "$lib/client";
+import { user } from "$lib/stores/userStore.svelte";
 import { translationIds } from "$lib/translations";
 
 export type Translation = typeof translationIds;
@@ -9,11 +10,15 @@ function createI18n() {
 	});
 	let currentLocale = $state("de");
 	let locales = $state(["de"] as Array<string>);
+	const idLocale = "[ids]";
 	return {
 		get tr(): Translation {
 			return translations[currentLocale];
 		},
 		get locale(): string {
+			return currentLocale === idLocale ? "de" : currentLocale;
+		},
+		get selectedLocale(): string {
 			return currentLocale;
 		},
 		set locale(locale: string) {
@@ -43,6 +48,18 @@ function createI18n() {
 						}
 					} catch {
 						console.log(`Failed to fetch locale ${locale}`);
+					}
+				}
+				if (user?.data?.is_superuser) {
+					// add ids to locales for admin users
+					locales.push(idLocale);
+					translations[idLocale] = {};
+					for (const [section_key, section] of Object.entries(translationIds)) {
+						translations[idLocale][section_key] = {};
+						for (const item_key of Object.keys(section)) {
+							translations[idLocale][section_key][item_key] =
+								`[${section_key}.${item_key}]`;
+						}
 					}
 				}
 			}
