+++ frontend/src/routes/+layout.svelte
@@ -55,14 +55,12 @@ let { children } = $props();
 let hideDrawer = $state(true);
 
 onMount(async () => {
-	await i18n.load();
 	await user.load();
+	await i18n.load();
 });
 function toggleDrawer() {
 	hideDrawer = !hideDrawer;
 }
-
-const asAlert = false;
 </script>
 
 <svelte:head>
