+++ frontend/src/routes/forgotPassword/+page.svelte
@@ -1,4 +1,6 @@
-<script>
+<svelte:options runes={true} />
+
+<script lang="ts">
 import UserForgotPassword from "$lib/components/UserForgotPassword.svelte";
 </script>
