+++ frontend/src/routes/login/+page.svelte
@@ -1,4 +1,6 @@
-<script>
+<svelte:options runes={true} />
+
+<script lang="ts">
 import UserLogin from "$lib/components/UserLogin.svelte";
 </script>
